[ADD] errors of only images for profile and messages
