+++ src/components/UserFormDetails/index.js
@@ -8,7 +8,6 @@ import { FormInput, ActionsForm, SkeletonForm } from './styles'
 
 import { Input } from '../../styles/Inputs'
 import { Button } from '../../styles/Buttons'
-import { Alert } from '../Confirm'
 import { InputPhoneNumber } from '../InputPhoneNumber'
 
 import { flatArray } from '../../utils'
@@ -28,25 +27,18 @@ export const UserFormDetails = (props) => {
     isRequiredField,
     validationFields,
     handleChangeInput,
-    handleButtonUpdateClick
+    handleButtonUpdateClick,
+    setAlertState
   } = props
 
   const { handleSubmit, register, errors } = useForm()
 
   const [{ user }] = useSession()
-  const [alertState, setAlertState] = useState({ open: false, content: [] })
   const [isValidPhoneNumber, setIsValidPhoneNumber] = useState(null)
   const [validationFieldsSorted, setValidationFieldsSorted] = useState([])
   const [userPhoneNumber, setUserPhoneNumber] = useState(null)
 
   const showInputPhoneNumber = () => validationFields?.fields?.cellphone?.enabled ?? false
-
-  const closeAlert = () => {
-    setAlertState({
-      open: false,
-      content: []
-    })
-  }
 
   const setUserCellPhone = (isEdit = false) => {
     if (userPhoneNumber && !userPhoneNumber.includes('null') && !isEdit) {
@@ -282,16 +274,6 @@ export const UserFormDetails = (props) => {
           </SkeletonForm>
         )}
       </FormInput>
-
-      <Alert
-        title={t('PROFILE', 'Profile')}
-        content={alertState.content}
-        acceptText={t('ACCEPT', 'Accept')}
-        open={alertState.open}
-        onClose={() => closeAlert()}
-        onAccept={() => closeAlert()}
-        closeOnBackdrop={false}
-      />
     </>
   )
 }
