+++ src/components/UserProfileForm/index.js
@@ -46,15 +46,27 @@ const UserProfileFormUI = (props) => {
   const [, t] = useLanguage()
   const [{ user }] = useSession()
   const [edit, setEdit] = useState(false)
-  const [openAlert, setOpenAlert] = useState(false)
+  const [alertState, setAlertState] = useState({ open: false, content: [] })
   const inputRef = useRef(null)
 
   const handleFiles = (files) => {
-    if (bytesConverter(inputRef.current?.files[0].size) > 2048) {
-      setOpenAlert(true)
-      return
-    }
     if (files.length === 1) {
+      const type = files[0].type.split('/')[0]
+      if (type !== 'image') {
+        setAlertState({
+          open: true,
+          content: [t('ERROR_ONLY_IMAGES', 'Only images can be accepted')]
+        })
+        return
+      }
+
+      if (bytesConverter(files[0].size) > 2048) {
+        setAlertState({
+          open: true,
+          content: [t('IMAGE_MAXIMUM_SIZE', 'The maximum image size is 2 megabytes')]
+        })
+        return
+      }
       handlechangeImage(files[0])
     }
   }
@@ -69,6 +81,13 @@ const UserProfileFormUI = (props) => {
 
   const handleClickImage = () => {
     inputRef.current.click()
+  }
+
+  const closeAlert = () => {
+    setAlertState({
+      open: false,
+      content: []
+    })
   }
 
   useEffect(() => {
@@ -143,6 +162,7 @@ const UserProfileFormUI = (props) => {
                   {...props}
                   onCancel={toggleEditState}
                   onCloseProfile={() => setEdit(false)}
+                  setAlertState={setAlertState}
                 />
               </WrapperForm>
             )}
@@ -155,12 +175,12 @@ const UserProfileFormUI = (props) => {
         </SavedPlaces>
       </Container>
       <Alert
-        title={t('USER_PROFILE', 'User Profile')}
-        content={t('IMAGE_MAXIMUM_SIZE', 'The maximum image size is 2 megabytes')}
+        title={t('PROFILE', 'Profile')}
+        content={alertState.content}
         acceptText={t('ACCEPT', 'Accept')}
-        open={openAlert}
-        onClose={() => setOpenAlert(false)}
-        onAccept={() => setOpenAlert(false)}
+        open={alertState.open}
+        onClose={() => closeAlert()}
+        onAccept={() => closeAlert()}
         closeOnBackdrop={false}
       />
     </>
